refactor: eliminate all clippy cast_possible_truncation warnings
Host-side crates get checked conversions: serializer counts/offsets go
through c16/c32 try_from helpers or per-site try_from().expect(); the
previously unguarded tzid-pool u16 offset table gets an ensure! guard
(Error::FormatLimit) matching the CSR loud-failure policy. Two real
display bugs fixed: rdp_sweep and size_table printed eps_m as u64,
truncating fractional epsilon.

The device crate (utz) instead uses fn-level #[expect(..., reason)]:
its casts are infallible by validated-format invariants and a checked
cast would add a per-vertex branch to the integer-bound PIP loop plus
panic strings to flash. Float->int sites everywhere use narrow
#[expect] with reasons since Rust has no fallible float conversion.

+++ utz-build/src/cmd/rdp_sweep.rs
@@ -45,7 +45,7 @@ pub fn run(a: Args) -> utz_build::Result<()> {
             if lookup(&srefs, lo, la) != truth[i] { miss += 1; }
         }
         println!("{:>8}{:>10}{:>8.1}%{:>12}{:>12}{:>12}{:>10.3}%",
-            eps_m as u64, out.verts, 100.0 * out.verts as f64 / v0 as f64,
+            eps_m, out.verts, 100.0 * out.verts as f64 / v0 as f64,
             raw, z, x, 100.0 * miss as f64 / sample as f64);
     }
     Ok(())
